Agregado FlushDB

+++ src/comu/ServicioEntidad.java
@@ -12,31 +12,31 @@ public abstract class ServicioEntidad<E> implements Serializable {
 	private static final long serialVersionUID = 1L;
 	protected EntityManagerFactory emf;
 	protected EntityManager em;
-	
+
 	public ServicioEntidad() {
 		super();
 	}
 
 	public void startEntityManager() {
-		if(emf == null){
-		emf = Persistence.createEntityManagerFactory("DogeBarcelo");
+		if (emf == null) {
+			emf = Persistence.createEntityManagerFactory("DogeBarcelo");
 		}
-		if(em == null){
-		em = emf.createEntityManager();
+		if (em == null) {
+			em = emf.createEntityManager();
 		}
 	}
 
-	public EntityManager getEntityManager(){
+	public EntityManager getEntityManager() {
 		return em;
 	}
-	
+
 	public void closeEntityManager() {
 		em.close();
 		emf.close();
 	}
 
 	public void insertar(E item) {
-		if(em == null){
+		if (em == null) {
 			startEntityManager();
 		}
 		em.getTransaction().begin();
@@ -46,7 +46,7 @@ public abstract class ServicioEntidad<E> implements Serializable {
 	}
 
 	public void actualizar(E item) {
-		if(em == null){
+		if (em == null) {
 			startEntityManager();
 		}
 		em.getTransaction().begin();
@@ -54,37 +54,49 @@ public abstract class ServicioEntidad<E> implements Serializable {
 		em.flush();
 		em.getTransaction().commit();
 	}
-	
+
 	public void eliminar(E item) {
-		if(em == null){
+		if (em == null) {
 			startEntityManager();
 		}
 		em.getTransaction().begin();
-		em.remove(item);
+		if (item.equals(null)) {
+			em.remove(item);
+		}
 		em.flush();
 		em.getTransaction().commit();
 	}
-	
-	/*###################################################################################*/
-	
-	/** En teoria estos metodos no se deben de usar, pero aqui se quedan para referencia :D */
-	public E buscar(E item, Object pk){
-		if(em == null){
+
+	/*
+	 * ##########################################################################
+	 * #########
+	 */
+
+	/**
+	 * En teoria estos metodos no se deben de usar, pero aqui se quedan para
+	 * referencia :D
+	 */
+	public E buscar(E item, Object pk) {
+		if (em == null) {
 			startEntityManager();
 		}
 		E pop = (E) em.find(item.getClass(), pk);
 		return pop;
 	}
 
-	public List<E> buscarTodos(E item){
-		if(em == null){
+	public List<E> buscarTodos(E item) {
+		if (em == null) {
 			startEntityManager();
 		}
-		String qjl = "Select t from " + item.getClass().getSimpleName() +" t";
-		List<E> popio = (List<E>) em.createQuery(qjl, item.getClass()).getResultList();
+		String qjl = "Select t from " + item.getClass().getSimpleName() + " t";
+		List<E> popio = (List<E>) em.createQuery(qjl, item.getClass())
+				.getResultList();
 		return popio;
 	}
 
-	/*###################################################################################*/
-	
+	/*
+	 * ##########################################################################
+	 * #########
+	 */
+
 }
